[TLC-50] Rewrite LDAPAuthentication
LDAPAuthentication contained a lot of bugs, especially when adding
EPersons to Groups "for the remainder of their current, logged in
session". Attributes like given and family names were not taken over
when the user's credentials where used for the LDAP search. As this
needs a lot of testing, we created a rewrite in a separate class.